third_party/WebKit/LayoutTests/fast/dom/custom/registration-context-delete-during-register-created-retrieval-expected.txt

CONSOLE WARNING: line 24: document.registerElement is deprecated and will be removed in M73, around March 2019. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 for more details.
Tests destroying a context during registration at the point when the created callback is retrieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS destroyed context
PASS caught expected exception "InvalidStateError: Failed to execute 'registerElement' on 'Document': Registration failed for type 'x-a'. The context is no longer valid."
PASS child done
PASS successfullyParsed is true

TEST COMPLETE
